Fix S-12 review comments: @MainActor on testPerformAllowMainThread, simplify testCancelWithNoWriterIsNoOp, extract shared writeSampleMovie helper
- AsyncBridgeTests: mark testPerformAllowMainThread @MainActor with Thread.isMainThread assertion.
- MovieMutatorTransformExportTests: use MovieMutator(with: AVMutableMovie()) for testCancelWithNoWriterIsNoOp.
- Extract writeSampleMovie into shared TestMovieFixtureWriter.swift, remove duplicate from both test files.
- MovieMutatorEditTests: update makeMutator call to use shared helper signature.
