+++ cutter2Tests/TestMovieFixtureWriter.swift
@@ -1,0 +1,79 @@
+//  TestMovieFixtureWriter.swift
+//  cutter2Tests
+//
+//  Shared test helper: write a sample H.264 .mov off the main thread.
+
+import AVFoundation
+import CoreMedia
+import Foundation
+
+/// nonisolated helper: write a sample H.264 .mov off the main thread
+func writeSampleMovie(
+    to url: URL,
+    duration: TimeInterval = 1.0,
+    timescale: CMTimeScale = 600,
+    frameRate: Int = 30
+) -> Bool {
+    guard let writer = try? AVAssetWriter(outputURL: url, fileType: .mov) else { return false }
+
+    let width = 320
+    let height = 180
+    let videoSettings: [String: Any] = [
+        AVVideoCodecKey: AVVideoCodecType.h264,
+        AVVideoWidthKey: width,
+        AVVideoHeightKey: height
+    ]
+    let input = AVAssetWriterInput(mediaType: .video, outputSettings: videoSettings)
+    input.expectsMediaDataInRealTime = false
+    let adaptor = AVAssetWriterInputPixelBufferAdaptor(
+        assetWriterInput: input,
+        sourcePixelBufferAttributes: [
+            kCVPixelBufferPixelFormatTypeKey as String: kCVPixelFormatType_32BGRA,
+            kCVPixelBufferWidthKey as String: width,
+            kCVPixelBufferHeightKey as String: height
+        ]
+    )
+    guard writer.canAdd(input) else { return false }
+    writer.add(input)
+
+    guard writer.startWriting() else { return false }
+    writer.startSession(atSourceTime: .zero)
+
+    let attrs: [CFString: Any] = [
+        kCVPixelBufferCGImageCompatibilityKey: true,
+        kCVPixelBufferCGBitmapContextCompatibilityKey: true,
+        kCVPixelBufferIOSurfacePropertiesKey: [:] as CFDictionary
+    ]
+    var pixelBuffer: CVPixelBuffer?
+    let pstatus = CVPixelBufferCreate(kCFAllocatorDefault, width, height,
+                                      kCVPixelFormatType_32BGRA, attrs as CFDictionary, &pixelBuffer)
+    guard pstatus == kCVReturnSuccess, let pb = pixelBuffer else { return false }
+    CVPixelBufferLockBaseAddress(pb, [])
+    if let base = CVPixelBufferGetBaseAddress(pb) {
+        memset(base, 0, CVPixelBufferGetDataSize(pb))
+    }
+    CVPixelBufferUnlockBaseAddress(pb, [])
+
+    let frameDuration = CMTime(value: CMTimeValue(timescale) / CMTimeValue(frameRate), timescale: timescale)
+    let frameCount = Int((duration * Double(frameRate)).rounded())
+    let maxReadySpins = 5_000 // 5s at 1ms
+    var presentation = CMTime.zero
+    for _ in 0..<frameCount {
+        var spins = 0
+        while !input.isReadyForMoreMediaData {
+            spins += 1
+            if spins > maxReadySpins { return false }
+            usleep(1000)
+        }
+        let ok = adaptor.append(pb, withPresentationTime: presentation)
+        if !ok { return false }
+        presentation = CMTimeAdd(presentation, frameDuration)
+    }
+    input.markAsFinished()
+    writer.endSession(atSourceTime: presentation)
+    let group = DispatchGroup()
+    group.enter()
+    writer.finishWriting { group.leave() }
+    group.wait()
+    return writer.status == .completed
+}
